HomePage를 StatelessWidget으로 변경하고, 애니메이션 로직을 챌린지 요구사항에 따라 PageFlipBuilder 클래스로 분리
